+++ src/include/tempus.hpp
@@ -19,14 +19,14 @@ namespace Tempus{
     template<typename TICK>
     using Time = std::chrono::time_point<Clock, TICK>;
 
-    typedef std::chrono::duration<unsigned long long, std::nano> Nanosecond;
-    typedef std::chrono::duration<long, std::micro> Microsecond;
-    typedef std::chrono::duration<long, std::milli> Millisecond;
-    typedef std::chrono::duration<long, std::ratio<1> > Second;
-    typedef std::chrono::duration<long, std::ratio<60> > Minute;
-    typedef std::chrono::duration<long, std::ratio<3600> > Hour;
-    typedef std::chrono::duration<long, std::ratio<86400> > Day;
-    typedef std::chrono::duration<long, std::ratio<31556952> > Year;
+    typedef std::chrono::duration<uint64_t, std::nano> Nanosecond;
+    typedef std::chrono::duration<int32_t, std::micro> Microsecond;
+    typedef std::chrono::duration<int32_t, std::milli> Millisecond;
+    typedef std::chrono::duration<int32_t, std::ratio<1> > Second;
+    typedef std::chrono::duration<int32_t, std::ratio<60> > Minute;
+    typedef std::chrono::duration<int32_t, std::ratio<3600> > Hour;
+    typedef std::chrono::duration<int32_t, std::ratio<86400> > Day;
+    typedef std::chrono::duration<int32_t, std::ratio<31556952> > Year;
 
     /*
         DURATION INSTANCES
@@ -47,48 +47,48 @@ namespace Tempus{
     */
 
     // Fractional seconds to seconds
-    constexpr unsigned long nsecs_per_sec = std::nano::den;
-    constexpr unsigned long usecs_per_sec = std::micro::den;
-    constexpr unsigned long msecs_per_sec = std::milli::den;
+    constexpr uint32_t nsecs_per_sec = std::nano::den;
+    constexpr uint32_t usecs_per_sec = std::micro::den;
+    constexpr uint32_t msecs_per_sec = std::milli::den;
 
     // Seconds => Minutes => Hours => Days => Year
-    constexpr unsigned long secs_per_min = 60;
-    constexpr unsigned long mins_per_hour = 60;
-    constexpr unsigned long hours_per_day = 24;
-    constexpr unsigned long days_per_year = 365;
+    constexpr uint32_t secs_per_min = 60;
+    constexpr uint32_t mins_per_hour = 60;
+    constexpr uint32_t hours_per_day = 24;
+    constexpr uint32_t days_per_year = 365;
 
     // Nanoseconds to larger units
-    constexpr unsigned long nsecs_per_usec = usecs_per_sec / nsecs_per_sec;
-    constexpr unsigned long nsecs_per_msec = msecs_per_sec / nsecs_per_sec;
-    constexpr unsigned long nsecs_per_min = nsecs_per_sec * secs_per_min;
-    constexpr unsigned long nsecs_per_hour = nsecs_per_min * mins_per_hour;
-    constexpr unsigned long nsecs_per_day = nsecs_per_hour * hours_per_day;
-    constexpr unsigned long nsecs_per_year = nsecs_per_day * days_per_year;
+    constexpr uint32_t nsecs_per_usec = usecs_per_sec / nsecs_per_sec;
+    constexpr uint32_t nsecs_per_msec = msecs_per_sec / nsecs_per_sec;
+    constexpr uint32_t nsecs_per_min = nsecs_per_sec * secs_per_min;
+    constexpr uint32_t nsecs_per_hour = nsecs_per_min * mins_per_hour;
+    constexpr uint32_t nsecs_per_day = nsecs_per_hour * hours_per_day;
+    constexpr uint32_t nsecs_per_year = nsecs_per_day * days_per_year;
 
     // Microseconds to larger units
-    constexpr unsigned long usecs_per_msec = msecs_per_sec / nsecs_per_sec;
-    constexpr unsigned long usecs_per_min = usecs_per_sec * secs_per_min;
-    constexpr unsigned long usecs_per_hour = usecs_per_min * mins_per_hour;
-    constexpr unsigned long usecs_per_day = usecs_per_hour * hours_per_day;
-    constexpr unsigned long usecs_per_year = usecs_per_day * days_per_year;
+    constexpr uint32_t usecs_per_msec = msecs_per_sec / nsecs_per_sec;
+    constexpr uint32_t usecs_per_min = usecs_per_sec * secs_per_min;
+    constexpr uint32_t usecs_per_hour = usecs_per_min * mins_per_hour;
+    constexpr uint32_t usecs_per_day = usecs_per_hour * hours_per_day;
+    constexpr uint32_t usecs_per_year = usecs_per_day * days_per_year;
 
     // Milliseconds to larger units
-    constexpr unsigned long msecs_per_min = msecs_per_sec * secs_per_min;
-    constexpr unsigned long msecs_per_hour = msecs_per_min * mins_per_hour;
-    constexpr unsigned long msecs_per_day = msecs_per_hour * hours_per_day;
-    constexpr unsigned long msecs_per_year = msecs_per_day * days_per_year;
+    constexpr uint32_t msecs_per_min = msecs_per_sec * secs_per_min;
+    constexpr uint32_t msecs_per_hour = msecs_per_min * mins_per_hour;
+    constexpr uint32_t msecs_per_day = msecs_per_hour * hours_per_day;
+    constexpr uint32_t msecs_per_year = msecs_per_day * days_per_year;
 
     // Seconds to larger units
-    constexpr unsigned long secs_per_hour = secs_per_min * mins_per_hour;
-    constexpr unsigned long secs_per_day = secs_per_hour * hours_per_day;
-    constexpr unsigned long secs_per_year = secs_per_day * days_per_year;
+    constexpr uint32_t secs_per_hour = secs_per_min * mins_per_hour;
+    constexpr uint32_t secs_per_day = secs_per_hour * hours_per_day;
+    constexpr uint32_t secs_per_year = secs_per_day * days_per_year;
 
     // Minutes to larger units
-    constexpr unsigned long mins_per_day = mins_per_hour * hours_per_day;
-    constexpr unsigned long mins_per_year = mins_per_day * days_per_year;
+    constexpr uint32_t mins_per_day = mins_per_hour * hours_per_day;
+    constexpr uint32_t mins_per_year = mins_per_day * days_per_year;
 
     // Hours to larger units
-    constexpr unsigned long hours_per_year = hours_per_day * days_per_year;
+    constexpr uint32_t hours_per_year = hours_per_day * days_per_year;
 
     /*
         FORMAT SPECS
@@ -124,7 +124,7 @@ namespace Tempus{
     template<typename TICK=Nanosecond>
     std::string strtime(const char *spec=tm_spec);
 
-    std::string strtime(unsigned long long t0, const char *spec=tm_spec);
+    std::string strtime(uint64_t t0, const char *spec=tm_spec);
 
     std::string strdate(const char *spec=dt_spec);
